Fixed compiler warnings.
git-svn-id: svn+ssh://svn.code.sf.net/p/supertuxkart/code/main/trunk@8699 178a84e3-b1eb-0310-8ba1-8eac791a3b58

+++ src/guiengine/widgets/dynamic_ribbon_widget.cpp
@@ -242,7 +242,7 @@ void DynamicRibbonWidget::add()
                 // will be penalized. A configuration that makes items much smaller than
                 // requested in the XML file will also be penalized.
                 float ratio = (float)item_height / (float)m_child_height;
-                float total_area = m_w * m_h;
+                float total_area = (float)(m_w * m_h);
                 const float score = log(2.0f*visible_items) *
                                       std::min(ratio, 1.0f) * std::min(taken_area/total_area, 1.0f);
